backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
pymongo
pinecone
boto3
python-dotenv
